requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

sqlalchemy==2.0.29
aiosqlite==0.20.0
alembic==1.13.1

argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

twilio==9.0.4
qrcode[pil]==7.4.2

reportlab==4.0.9
weasyprint==60.1
pypdf2==3.0.1

requests==2.32.3

pytest==7.4.4
pytest-asyncio==0.23.6
httpx==0.27.0
black==24.4.2
flake8==7.0.0
mypy==1.9.0
